requirements.txt
flake8
ttkbootstrap
ttkthemes
pillow
pyinstaller